Add @MetaMask/delegation as co-owners of permission signature confirmation code - move typed-sign-permission.tsx (and tests) into typed-sign-permission directory for increased cohesion
