+++ ui/pages/confirmations/components/confirm/info/typed-sign/typed-sign-permission/index.ts
@@ -1,0 +1,1 @@
+export { default } from './typed-sign-permission';
