+++ ui/pages/confirmations/components/confirm/info/typed-sign/typed-sign-permission/typed-sign-permission.test.tsx
@@ -1,16 +1,13 @@
 import { DecodedPermission } from '@metamask/gator-permissions-controller';
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
-import { getMockTypedSignPermissionConfirmState } from '../../../../../../../test/data/confirmations/helper';
-import {
-  renderWithConfirmContextProvider,
-  renderWithConfirmContext,
-} from '../../../../../../../test/lib/confirmations/render-helpers';
-import { enLocale as messages } from '../../../../../../../test/lib/i18n-helpers';
+import { getMockTypedSignPermissionConfirmState } from '../../../../../../../../test/data/confirmations/helper';
+import { renderWithConfirmContextProvider, renderWithConfirmContext } from '../../../../../../../../test/lib/confirmations/render-helpers';
+import { enLocale as messages } from '../../../../../../../../test/lib/i18n-helpers';
 import TypedSignPermissionInfo from './typed-sign-permission';
 
 jest.mock(
-  '../../../../../../components/app/alert-system/contexts/alertMetricsContext',
+  '../../../../../../../components/app/alert-system/contexts/alertMetricsContext',
   () => ({
     useAlertMetrics: jest.fn(() => ({
       trackAlertMetrics: jest.fn(),
@@ -18,8 +15,8 @@ jest.mock(
   }),
 );
 
-jest.mock('../../../../utils/token', () => ({
-  ...jest.requireActual('../../../../utils/token'),
+jest.mock('../../../../..//utils/token', () => ({
+  ...jest.requireActual('../../../../../utils/token'),
   fetchErc20DecimalsOrThrow: jest.fn().mockResolvedValue(18),
 }));
